backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
supabase==2.18.1
python-dotenv==1.1.1
httpx==0.28.1
PyJWT==2.10.1
